requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
requests>=2.28.0
google-generativeai>=0.3.0
twilio>=8.0.0
sqlalchemy>=2.0.0
